requirements/development.txt
ipython
flake8
django-extensions
